Add funcs to determine if segment2 and point2 are collinear
Add multiple overloads for different combinations of point2 and segment2. There must be at least three point2 if there are no segment2, but in all other cases there must be two shapes as args. Add an overloaded version that accepts a dynamic number of point2 and segment2 using recursive templates. Add a concept that includes only segment2 and point2 to help define this template.